port Sean Luke's Flockers demo
It is complete in the sense that it reproduces the intended simulation with
a nice visual extension. It also shows how Scala can make things more clear
than Java; however, the component-style design that I am after was not
implemented due to time limitations. After I port SugarScape, I'll finish
this demo.